+++ apm-agent-plugins/apm-jsf-plugin/src/main/java/co/elastic/apm/agent/jsf/JakartaeeJsfLifecycleRenderInstrumentation.java
@@ -1,0 +1,13 @@
+package co.elastic.apm.agent.jsf;
+
+public class JakartaeeJsfLifecycleRenderInstrumentation extends AbstractJsfLifecycleRenderInstrumentation {
+    @Override
+    String lifecycleClassName() {
+        return "jakarta.faces.lifecycle.Lifecycle";
+    }
+
+    @Override
+    String facesContextClassName() {
+        return "jakarta.faces.context.FacesContext";
+    }
+}
